D9 p1

See merge request uh2r78/advent2019!6

+++ src/test/java/utils/MachineLauncher.java
@@ -10,8 +10,8 @@ import java.io.IOException;
 import java.util.List;
 
 public interface MachineLauncher {
-    default List<Integer> runMachineAndGetEndList(String path) throws IOException, PointeurException, CodeInconnuException, ModeInconnuException {
-        List<Integer> integerList = Parsers.recupererListeEntiers(path);
+    default List<Long> runMachineAndGetEndList(String path) throws IOException, PointeurException, CodeInconnuException, ModeInconnuException {
+        List<Long> integerList = Parsers.recupererListeEntiers(path);
         Machine machine = new Machine(integerList);
         machine.execute();
         return integerList;
